+++ extensions/CHANGELOG.md
@@ -21,7 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Renamed `parseLockfile` to `PhylumApi::parseDependencyFile`
 - Removed `lockfile` field from `PhylumApi::Package` type
 - Removed `PhylumApi::Lockfile` type in favor of `PhylumApi::DependencyFile`
-- Changed `PhylumApi::analyze` packages type `PhylumApi::PackageWithOrigin`
+- Changed `PhylumApi::analyze` packages type to `PhylumApi::PackageWithOrigin`
 
 ### Fixed