+++ extensions/CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Accept PURLs in `PhylumApi::analyze`
+
 ## 6.0.0 - 2023-12-13
 
 ### Changed